feat: add auth options to appBar
