+++ src/layouts/AuthLayout.tsx
@@ -1,12 +1,10 @@
-import { useCallback, useEffect} from "react";
+import { useCallback, useEffect } from "react";
 import useAuthStore from "../stores/AuthStore";
 
 export default function AuthLayout({ children }: { children: React.ReactNode }) {
 
   const {
-    isAuthenticated,
     setIsAuthenticated,
-    user,
     setUser
   } = useAuthStore();
 
@@ -31,24 +29,8 @@ export default function AuthLayout({ children }: { children: React.ReactNode }) 
     checkAuthStatus();
   }, [checkAuthStatus])
 
-  const handleLogin = () => {
-    // posible manera de redirección, requiere configuración en el servidor
-    // window.location.href = 'http://localhost:4006/api/auth/login?redirect_uri=http://localhost:3000/auth/callback';
-    window.location.href = 'http://localhost:4006/api/auth/login';
-  }
-
-  const handleLogout = async () => {
-    window.location.href = 'http://localhost:4006/api/auth/logout';
-  }
-
   return (
     <>
-      {user && <p>{user.uCorreo}</p>}
-      {
-        isAuthenticated
-          ? (<button onClick={handleLogout}>Logout</button>)
-          : (<button onClick={handleLogin}>Login</button>)
-      }
       {children}
     </>
   )
